Switched DBFacade and CheepService methods arround
