+++ src/Chirp.Razor/CheepService/ICheepService.cs
@@ -4,7 +4,7 @@ using DataTransferClasses;
 //REPO is synonymous with Service, but that name is taken. Should be changed to service when the other is deleted
 public interface ICheepService 
 {
-    ChirpDBContext context { get; set; }
+    DBFacade context { get; set; }
     public List<CheepDTO> GetCheeps(int page);
 
     public List<CheepDTO> GetCheepsFromAuthor(string author, int page);
